+++ NEWS.md
@@ -4,9 +4,13 @@
 ## Bug fixes:
 * Fixed bug in `abs_cat_stats` function to avoid downloading the files multiple
   times and apply `abs_cat_unzip` function only to compressed files.
+* Added OpenXml Spreadsheet files to list of valid ABS catalogue statistics file
+  types.
+  
 
 ## Changes:
 * Added `na.rm` argument to `abs_cat_stats` and `abs_read_tss` functions.
+* Added `abs_cat_releases` function.
 
 
 # raustats 0.1.0